cron_app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.3.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
audioread==3.0.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.3.6
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
cryptography==43.0.1
datasets==2.14.0
decorator==5.2.1
dill==0.3.7
fastapi==0.104.1
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.0.2
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
gunicorn==22.0.0
h11==0.16.0
hf-xet==1.1.5
httptools==0.6.4
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.5.4
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.44.0
m3u8==6.0.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.15
mutagen==1.47.0
networkx==3.5
numba==0.61.2
numpy==1.24.3
openai-whisper==20250625
packaging==25.0
pandas==2.3.1
parso==0.8.4
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg==3.2.9
pyarrow==20.0.0
PyAudio==0.2.14
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.5.0
pydantic_core==2.14.1
pydub==0.25.1
pyOpenSSL==24.2.1
pyparsing==3.1.4
PyQt5==5.15.10
PyQt5-Qt5==5.15.13
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
QScintilla==2.14.1
redis==5.0.6
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.18.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
srt==3.5.3
starlette==0.47.1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.15.2
toolz==0.12.1
torch==2.1.0
torchaudio==2.1.0
tqdm==4.67.1
transformers==4.36.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Unidecode==1.3.8
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
vine==5.1.0
vosk==0.3.44
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
Yapsy==1.12.2
yarl==1.20.1
yt-dlp==2025.6.30